requirements.txt
azure-core==1.32.0
azure-cosmos==4.9.0
azure-search-documents==11.5.2
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.2.14